fix(cli): keep load-boundary schemas free of the core dependency
The doc/template Zod schemas are imported by CLI code on the install-time
hot path (template + doc discovery, reached by `astryx init`, which some
apps run from a postinstall). Co-locating them with the factory re-export
made loading a schema transitively resolve `@astryxdesign/core/authoring`
-> core's built `dist/`, which does not exist yet during install — breaking
`astryx init` and every downstream job.

Move the schemas into core-free modules (`src/schemas/{template,doc}-schema.mjs`)
and point the internal importers (`api/template.mjs`, `lib/component-loader.mjs`)
at them directly. The public `@astryxdesign/cli/{template,doc}` subpaths still
re-export both the factories (from core) and the schemas, so their surface is
unchanged. Only the authoring subpaths pull core now, and those are only
imported after core is built.
